pyproject.toml
[project]
name = "base24-base16"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorzero>=2.0",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
